fixing

+++ src/mappings/swaps.ts
@@ -119,7 +119,7 @@ export function handleSwap(event: Swap): void {
         candle.lastBlock = event.block.number.toI32();
         candle.token0TotalAmount = candle.token0TotalAmount.plus(token0Amount);
         candle.token1TotalAmount = candle.token1TotalAmount.plus(token1Amount);
-        candle.tradeAmountUSD = candle.token1TotalAmount.toBigDecimal().times(bundle.bnbPrice);
+        candle.tradeAmountUSD = candle.token1TotalAmount.times(bundle.bnbPrice);
         candle.save();
     }
 }
